Retry requests in case of errors up to sendAttempts times
Instead of just reconnecting upon read errors, it is important to retry
the request as well. A common read error is EOF, which indicates that
the TCP connection is closed.

In case of read or write errors, retry sending the request up to
sendAttempts times. Monitor errors using the purged_tcp_errors_total
prometheus metric.

Bug: T249583
Change-Id: I5671b2f8be6fd0e96f27f5da242d3f9b2597215e