Add the ability to consume from kafka
We want to move our purge delivery method to kafka. Therefore, add an
optional reader that consumes kafka topics in addition to listening to
multicast messages. The purges obtained in this way will be relayed to
the same channel.

The code consumes events that conform to the resource_change wikimedia
event schema.

Bug: T133821
Change-Id: I408855c1ae28da38fcc67d25f25291ab4f08e82b